fix bug

+++ plot/net_plotter.py
@@ -37,7 +37,7 @@ def set_weights(net, weights, directions=None, step=None):
             changes = [d*step for d in directions[0]]
 
         for (p, w, d) in zip(net.parameters(), weights, changes):
-            p.data = w + torch.Tensor(d).type_as(w.dtype).cuda() # fix fp 16 bug
+            p.data = w + torch.Tensor(d).type_as(w).cuda() # fix fp 16 bug
 
 
 def set_states(net, states, directions=None, step=None):
